@@ -9,51 +9,53 @@
     "deploy": "gatsby build --prefix-paths && gh-pages -r \"https://${GITHUB_TOKEN}@github.com/ERS-HCL/ers-hcl.github.io.git\" -d public -b master"
   },
   "dependencies": {
-    "@material-ui/core": "^1.4.0",
-    "@material-ui/icons": "^1.1.0",
-    "babel-polyfill": "^6.26.0",
+    "@babel/polyfill": "^7.0.0",
+    "@material-ui/core": "^3.3.1",
+    "@material-ui/icons": "^3.0.1",
     "classnames": "^2.2.5",
+    "core-js": "^2.5.7",
     "dotenv": "^6.0.0",
-    "firebase": "^5.0.2",
-    "gatsby": "^1.9.243",
-    "gatsby-link": "^1.6.37",
-    "gatsby-paginate": "^1.0.12",
-    "gatsby-plugin-google-analytics": "^1.0.31",
+    "firebase": "^5.5.1",
+    "gatsby": "^2.0.31",
+    "gatsby-paginate": "^1.0.16",
+    "gatsby-plugin-google-analytics": "^2.0.6",
     "gatsby-plugin-google-fonts": "0.0.4",
-    "gatsby-plugin-guess-js": "^1.0.0-alpha.2",
-    "gatsby-plugin-manifest": "^1.0.15",
-    "gatsby-plugin-offline": "^1.0.15",
-    "gatsby-plugin-react-helmet": "^2.0.8",
-    "gatsby-source-filesystem": "^1.5.20",
+    "gatsby-plugin-guess-js": "^1.0.1",
+    "gatsby-plugin-manifest": "^2.0.6",
+    "gatsby-plugin-offline": "^2.0.9",
+    "gatsby-plugin-postcss": "^2.0.0",
+    "gatsby-plugin-react-helmet": "^3.0.0",
+    "gatsby-source-filesystem": "^2.0.5",
     "gatsby-source-github-api": "0.0.3",
-    "gatsby-transformer-excel": "^1.0.4",
-    "gatsby-transformer-remark": "^1.7.32",
-    "guess-webpack": "^0.1.0",
+    "gatsby-transformer-excel": "^2.1.2",
+    "gatsby-transformer-remark": "^2.1.9",
+    "guess-webpack": "^0.1.6",
     "immutable": "^3.8.2",
-    "lodash": "^4.17.5",
-    "moment": "^2.20.1",
-    "moment-timezone": "^0.5.14",
-    "react": "^16.3.2",
-    "react-csv": "^1.0.14",
-    "react-dom": "^16.3.2",
+    "lodash": "^4.17.11",
+    "moment": "^2.22.2",
+    "moment-timezone": "^0.5.21",
+    "postcss-browser-reporter": "^0.5.0",
+    "postcss-import": "^12.0.0",
+    "postcss-preset-env": "^6.0.6",
+    "postcss-reporter": "^6.0.0",
+    "react": "^16.5.2",
+    "react-csv": "^1.0.18",
+    "react-dom": "^16.5.2",
     "react-github-fork-ribbon": "^0.5.1",
     "react-helmet": "^5.2.0",
     "react-html-parser": "^2.0.2",
     "react-responsive": "^5.0.0",
-    "react-router-dom": "^4.2.2",
-    "react-table": "^6.7.6",
-    "react-tap-event-plugin": "^3.0.2",
+    "react-router-dom": "^4.3.1",
+    "react-table": "^6.8.6",
     "typeface-roboto": "0.0.54",
-    "typography": "^0.16.6",
-    "typography-theme-github": "^0.15.10",
-    "typography-theme-kirkham": "^0.16.3",
-    "uuid": "^3.2.1",
+    "uuid": "^3.3.2",
     "victory": "^0.26.0"
   },
   "devDependencies": {
-    "gh-pages": "^1.1.0"
+    "gh-pages": "^1.1.0",
+    "null-loader": "^0.1.1"
   },
-  "version": "1.9.6",
+  "version": "2.0.0",
   "main": "n/a",
   "repository": {
     "type": "git",
