@@ -1,25 +1,27 @@
-const path = require(`path`);
-const { createFilePath } = require(`gatsby-source-filesystem`);
-const createPaginatedPages = require('gatsby-paginate');
+const path = require(`path`)
+const { createFilePath } = require(`gatsby-source-filesystem`)
+const createPaginatedPages = require('gatsby-paginate')
 
-exports.onCreateNode = ({ node, getNode, boundActionCreators }) => {
-  const { createNodeField } = boundActionCreators;
+exports.onCreateNode = ({ node, getNode, actions }) => {
+  const { createNodeField } = actions
   if (node.internal.type === `MarkdownRemark`) {
-    const slug = createFilePath({ node, getNode, basePath: `pages` });
+    const slug = createFilePath({ node, getNode, basePath: `pages` })
     createNodeField({
       node,
       name: `slug`,
       value: slug
-    });
+    })
   }
-};
+}
 
-exports.createPages = ({ graphql, boundActionCreators }) => {
-  const { createPage } = boundActionCreators;
+exports.createPages = ({ graphql, actions }) => {
+  const { createPage } = actions
   return new Promise((resolve, reject) => {
     graphql(`
       {
-        posts: allMarkdownRemark(sort: { fields: [frontmatter___date], order: DESC }) {
+        posts: allMarkdownRemark(
+          sort: { fields: [frontmatter___date], order: DESC }
+        ) {
           totalCount
           edges {
             node {
@@ -45,7 +47,7 @@ exports.createPages = ({ graphql, boundActionCreators }) => {
         pageLength: 5, // This is optional and defaults to 10 if not used
         pathPrefix: 'blogs', // This is optional and defaults to an empty string if not used
         context: {} // This is optional and defaults to an empty object if not used
-      });
+      })
       result.data.posts.edges.map(({ node }) => {
         createPage({
           path: node.fields.slug,
@@ -54,9 +56,9 @@ exports.createPages = ({ graphql, boundActionCreators }) => {
             // Data passed to context is available in page queries as GraphQL variables.
             slug: node.fields.slug
           }
-        });
-      });
-      resolve();
-    });
-  });
-};
+        })
+      })
+      resolve()
+    })
+  })
+}
