@@ -1,8 +1,12 @@
-import 'babel-polyfill';
-import React from 'react';
+import 'core-js/es6/promise'
+import 'core-js/es6'
+import React from 'react'
 import Challenges from '../components/challenges/challenges'
+import Layout from '../components/layout'
 export default ({ data }) => {
   return (
-    <Challenges />
-  );
-};
+    <Layout>
+      <Challenges />
+    </Layout>
+  )
+}
