@@ -1,10 +1,31 @@
-import 'babel-polyfill'
+import 'core-js/es6/promise'
+import 'core-js/es6/symbol'
+import 'core-js/es6/object'
+import 'core-js/es6/function'
+import 'core-js/es6/parse-int'
+import 'core-js/es6/parse-float'
+import 'core-js/es6/number'
+import 'core-js/es6/math'
+import 'core-js/es6/string'
+import 'core-js/es6/date'
+import 'core-js/es6/array'
+import 'core-js/es6/regexp'
+import 'core-js/es6/map'
+import 'core-js/es6/weak-map'
+import 'core-js/es6/set'
 import React from 'react'
+import { graphql } from 'gatsby'
 import RepoList from '../components/repoList/repoList'
+import Layout from '../components/layout'
 
 export default ({ data }) => {
   return (
-    <RepoList githubData={data.allGithubData} buildTime={data.site.buildTime} />
+    <Layout>
+      <RepoList
+        githubData={data.allGithubData}
+        buildTime={data.site.buildTime}
+      />
+    </Layout>
   )
 }
 
