@@ -1,7 +1,6 @@
-import { ref, firebaseAuth } from './firebase'
+import { firebaseAuth } from './firebase'
 
 export function logout() {
-    
   return firebaseAuth().signOut()
 }
 
@@ -10,10 +9,10 @@ export function login(email, pw) {
 }
 
 export function loginOAuth() {
-  const provider = new firebaseAuth.GithubAuthProvider();
-  return firebaseAuth().signInWithPopup(provider);
+  const provider = new firebaseAuth.GithubAuthProvider()
+  return firebaseAuth().signInWithPopup(provider)
 }
 
-export function createUser (email, pw) {
+export function createUser(email, pw) {
   return firebaseAuth().createUserWithEmailAndPassword(email, pw)
 }
